+++ De la Puente, Matias/ejercicio 1/api/app.js
@@ -1,7 +1,9 @@
 import express from "express";
 import cors from "cors";
 import sumasRouter from "./sumas.js";
-import { divisionesRoute } from "./divisiones.js";
+import  restasRouter  from "./restas.js";
+import divisionesRouter from "./divisiones.js";
+import multiplicacionesRouter from "./multiplicaciones.js";
 
 const app = express();
 const port = 3000;
@@ -13,11 +15,14 @@ app.use(express.json());
 app.use(cors());
 
 app.get("/", (req, res) => {
-  res.send("Resultados");
+  res.send("Api en Funcionamiento");
 });
 
+// Rutas de operaciones
 app.use("/sumas", sumasRouter);
-app.use("/divisiones", divisionesRoute);
+app.use("/restas", restasRouter)
+app.use("/divisiones", divisionesRouter);
+app.use("/multiplicaciones", multiplicacionesRouter);
 
 app.listen(port, () => {
   console.log(`La aplicacion esta funcionando en: ${port}`);
