+++ De la Puente, Matias/ejercicio 1/api/restas.js
@@ -1,11 +1,15 @@
-import express from 'express'
+import express from "express";
 
 const router = express.Router();
 
-//cambio de signo (/)
+//cambio de signo a (-)
 
-let sumas = [];
-let sumasMaxId = sumas.length > 0 ? Math.max(...restas.map((r) => r.id)) : 0;
+let sumas = [
+  // { id: 1, a: 2, b: 5, resultado: 7 },
+  // { id: 2, a: 6, b: 81, resultado: 87 },
+  // { id: 5, a: 12, b: 55, resultado: 87 },
+];
+let sumasMaxId = 0;
 
 // GET /sumas
 router.get("/", (req, res) => {
@@ -27,8 +31,7 @@ router.get("/:id", (req, res) => {
 // POST /sumas
 router.post("/", (req, res) => {
   const { a, b } = req.body;
-  if (b === 0) { return res.status(400).send({mensaje: "No se divide por 0"})}
-  const suma = { id: ++sumasMaxId, a, b, resultado: a / b, fecha: new Date() };
+  const suma = { id: ++sumasMaxId, a, b, resultado: a - b, fecha: new Date() };
   sumas.push(suma);
   res.status(201).send({ suma });
 });
@@ -37,9 +40,24 @@ router.post("/", (req, res) => {
 router.put("/:id", (req, res) => {
   const id = parseInt(req.params.id);
   const { a, b } = req.body;
-
-  const sumaModificada = { id, a, b, resultado: a / b, fecha: new Date() };
-
+  // Con find
+  /*
+  const suma = sumas.find((suma) => suma.id === id);
+  suma.a = a;
+  suma.b = b;
+  suma.resultado = a + b;
+  suma.fecha = new Date();
+  */
+  const sumaModificada = { id, a, b, resultado: a - b, fecha: new Date() };
+  // con forEach
+  /*
+  sumas.forEach((suma, index) => {
+    if (suma.id === id) {
+      sumas[index] = sumaModificada;
+    }
+  });
+  */
+  // con map
   sumas = sumas.map((suma) => (suma.id === id ? sumaModificada : suma));
   res.status(200).send({ suma: sumaModificada });
 });
@@ -50,7 +68,5 @@ router.delete("/:id", (req, res) => {
   sumas = sumas.filter((suma) => suma.id !== id);
   res.status(200).send({ id });
 });
-
-
 
 export default router;
